Proposed changes to vlan-aware-vms.
This patch includes some proposed changers to the vlan-aware-vms spec.
I'd like to propose that we revisit the idea of using only the existing
Neutron port resource instead of creating a new trunk-port resource.

I have attempted to capture it in updates to the spec itself.  In short
though, it seems to me that using the existing port resource is
sufficient.  I believe it will result in a simpler implementation (though
that will have to be proven with code).  Finally, it doesn't require any
changes to projects that interface with Neutron and handle VIF port
binding (like Nova, but also Kuryr or anything similar).

Change-Id: I8563e54cc984b60f2be2eb7cc2e3c6fc60e6de20
Signed-off-by: Russell Bryant <rbryant@redhat.com>